added support for tag combinations
multiple tags are always combined with AND - at least until there's a
compelling use case that requires otherwise (in which case something
like whoosher might be more appropriate)

not sure whether it's a great idea to have this in the path rather than
as query parameters (cf. CSV parsing): seems friendlier, but leads to
ambiguity (multiple URIs for the same result) - @cdent?